Decouple VM debug symbols/names from opcodes.ts
`opcodes.ts` previously imported `@glimmer/debug` (DebugLogger,
VmSnapshot, debugOp / describeOp, opcodeMetadata, frag, etc.) at the
top level and assembled the per-opcode `debugBefore`/`debugAfter`
hooks inline in `AppendOpcodes`'s constructor — gated by `LOCAL_DEBUG`,
so dead in production, but the imports still pulled the heavy
`@glimmer/debug` graph into the bundle.

Same registration pattern as the DebugRenderTree split: opcodes.ts
exposes `registerDebugOpcodeSetup(setup)`; the heavy hook
implementation moved to `opcodes-debug-setup.ts`, which calls the
registry on import. `externs(vm)` now also requires the hooks to be
registered (returns `undefined` otherwise) so dev builds that don't
opt in skip the debug path entirely instead of crashing on a non-null
assertion.

Production hello-world holds at 134.12 KB / 42.90 KB gzip (`LOCAL_DEBUG`
already eliminated the hooks there); the analysis bundle drops the
`@glimmer/debug` files entirely.

Verified: lint clean (after `pnpm lint:fix`), type-checks pass,
hello-world builds, classic v2-app-template `pnpm test` 1/1 pass.
